+++ src/math/light_and_shading/set_material.c
@@ -12,21 +12,21 @@
 
 #include "minirt.h"
 
-static t_bool	is_not_default(float a);
-static void		init_phong_pointers(float *phong_values[4], t_material *m);
-static void		init_phong_values(float *phong_values[4], t_tuple new_phong);
+static t_bool	is_not_default(double a);
+static void		init_phong_pointers(double *phong_values[4], t_material *m);
+static void		init_phong_values(double *phong_values[4], t_tuple new_phong);
 static void		init_color_values(t_material *m, t_tuple new_color);
 
 void	set_material(t_tuple new_phong, t_tuple new_color, t_material *m)
 {
-	float	*phong_values[4];
+	double	*phong_values[4];
 
 	init_phong_pointers(phong_values, m);
 	init_phong_values(phong_values, new_phong);
 	init_color_values(m, new_color);
 }
 
-static void	init_phong_pointers(float *phong_values[4], t_material *m)
+static void	init_phong_pointers(double *phong_values[4], t_material *m)
 {
 	phong_values[0] = &m->ambient;
 	phong_values[1] = &m->diffuse;
@@ -34,7 +34,7 @@ static void	init_phong_pointers(float *phong_values[4], t_material *m)
 	phong_values[3] = &m->shininess;
 }
 
-static void	init_phong_values(float *phong_values[4], t_tuple new_phong)
+static void	init_phong_values(double *phong_values[4], t_tuple new_phong)
 {
 	int	counter;
 
@@ -58,7 +58,7 @@ static void	init_color_values(t_material *m, t_tuple new_color)
 	}
 }
 
-static t_bool	is_not_default(float a)
+static t_bool	is_not_default(double a)
 {
 	if (a != DEFAULT)
 		return (TRUE);
@@ -81,14 +81,14 @@ static t_bool	is_not_default(float a)
 // 	printf("shininess: %f\n", s.material.shininess);
 // 	print_tuple(s.material.color);
 //
-// 	cr_expect_eq(TRUE, floats_eq(s.material.ambient, 0.1));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.diffuse, 0.2));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.specular, 0.3));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.shininess, 0.4));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.color[R], 0.5));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.color[G], 0.6));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.color[B], 0.7));
-// 	cr_expect_eq(TRUE, floats_eq(s.material.color[W], 0.8));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.ambient, 0.1));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.diffuse, 0.2));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.specular, 0.3));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.shininess, 0.4));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.color[R], 0.5));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.color[G], 0.6));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.color[B], 0.7));
+// 	cr_expect_eq(TRUE, doubles_eq(s.material.color[W], 0.8));
 // }
 //
 // Test(test, set_material_default_untouched)
