+++ src/math/matrices/minors.c
@@ -12,12 +12,12 @@
 
 #include "minirt.h"
 
-float	_3x3minor(const t_3x3matrix m, int row, int col)
+double	_3x3minor(const t_3x3matrix m, int row, int col)
 {
 	return (_2x2determinant(_3x3submatrix(m, row, col)));
 }
 
-float	_4x4minor(const t_matrix m, int row, int col)
+double	_4x4minor(const t_matrix m, int row, int col)
 {
 	return (_3x3determinant(_4x4submatrix(m, row, col)));
 }
