+++ src/math/matrix_transformations/scaling.c
@@ -12,7 +12,7 @@
 
 #include "minirt.h"
 
-t_matrix	create_scaling_matrix(const float x, const float y, const float z)
+t_matrix	create_scaling_matrix(const double x, const double y, const double z)
 {
 	return (create_4x4_matrix(&(t_matrix){
 			.row_1 = {x, 0, 0, 0},
