+++ src/math/plane/intersect_plane.c
@@ -18,7 +18,7 @@ t_intersection	intersect_plane(t_shape **obj, t_tuple obj_dist_to_ray, t_ray r)
 
 	(void)obj_dist_to_ray;
 	ft_bzero((void *)&xs, sizeof(t_intersection));
-	if (fabsf(r.direction[Y]) < EPSILON)
+	if (fabs(r.direction[Y]) < EPSILON)
 		return (xs);
 	xs.head = intersection((-r.origin[Y] / r.direction[Y]), obj);
 	xs.count = 1;
