+++ src/math/rays/point.c
@@ -12,7 +12,7 @@
 
 #include "minirt.h"
 
-void	get_position(t_ray ray, float t, t_tuple result)
+void	get_position(t_ray ray, double t, t_tuple result)
 {
 	t_tuple	distance;
